db/etl: trim the fill-hint change
Build the hint map in NewAllocator instead of lazily in rememberFill, flatten
the Close block, and move the tests and the benchmark into etl_test.go and
etl_bench_test.go. Drops the churn test, which asserted what the bound test
already asserts, and the bench's smaller entry count.
